+++ packages/forms/examples/src/forms/pages/autocomplete/autocomplete.page.ts
@@ -100,6 +100,8 @@ public formatLabel(input: any) {
 		{name: 'Iron Man'},
 		{name: 'Deadpool'},
 	];
+	public disabledModel = 'Batman';
+	public isDisabled = true;
 	public searchValue = '';
 	public selectedItem1 = '';
 	public selectedItem2 = '';