+++ packages/forms/examples/src/forms/pages/datepicker/datepicker.page.ts
@@ -1,12 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 
 import { DateRange } from '@acpaas-ui/js-date-utils';
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 
 @Component({
 	templateUrl: './datepicker.page.html',
 })
-export class FormsDatepickerDemoPageComponent implements OnInit {
+export class FormsDatepickerDemoPageComponent implements OnInit, OnDestroy {
 	constructor(private fb: FormBuilder) { }
 
 	public dateRange: DateRange = [5, 6];
@@ -58,9 +60,29 @@ this.dateForm = this.fb.group({
 	</div>
 </form>`;
 
+	private destroyed$ = new Subject<boolean>();
+
 	public ngOnInit() {
 		this.dateForm = this.fb.group({
 			inputDate: ['07/10/2019'],
+			isDisabled: false,
 		});
+
+		this.dateForm.get('isDisabled').valueChanges
+			.pipe(
+				takeUntil(this.destroyed$),
+			)
+			.subscribe((isDisabled: boolean) => {
+				if (isDisabled) {
+					this.dateForm.get('inputDate').disable();
+				} else {
+					this.dateForm.get('inputDate').enable();
+				}
+			});
+	}
+
+	public ngOnDestroy() {
+		this.destroyed$.next(true);
+		this.destroyed$.complete();
 	}
 }