+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
+### Added
+- `forms`: add disabled state to all controls (except upload)
 
 ### Fixed
 - `autocomplete` Fixed  issues with items with identical labels and a custom template.